Update spec helpers
