+++ spec/support/helpers.rb
@@ -2,38 +2,35 @@ module Shore
   module Client
     module RSpec
       module Helpers
-        def merchant_role(attributes = {})
+        def urlsafe_uuid(uuid)
+          UUID.to_urlsafe(uuid)
+        end
+
+        def admin_merchant(id=nil)
           {
-            id: '74eb402b-e159-4027-9363-60772e6e8930',
-            type: 'merchants',
-            slug: 'achsel-alex',
-            name: 'Achsel Alex',
+            id: id || SecureRandom.uuid,
+            role: 'admin'
+          }
+        end
+
+        def owner_merchant(id=nil)
+          {
+            id: id || SecureRandom.uuid,
+            role: 'owner'
+          }
+        end
+
+        def member_merchant(id=nil)
+          {
+            id: id || SecureRandom.uuid,
             role: 'member'
-          }.with_indifferent_access.deep_merge(attributes)
+          }
         end
 
-        def merchant_account(attributes = {}, roles = [merchant_role])
-          {
-            'id' => '226fc766-3cf0-4d18-a988-5f8235f17edb',
-            'type' => 'merchant-accounts',
-            'attributes' => {
-              'name' => 'Bob Barker',
-              'roles' => roles
-            }
-          }.with_indifferent_access.deep_merge(attributes)
-        end
-
-        def jwt_payload(exp: (Time.now.utc + 2.days).beginning_of_day,
-                        member_roles: [member_role],
-                        attributes: {})
-          { exp: exp.to_i,
-            data: {
-              id: '226fc766-3cf0-4d18-a988-5f8235f17edb',
-              type: 'merchant-accounts',
-              attributes: {
-                name: 'Bob Barker',
-                roles: member_roles } }
-          }.with_indifferent_access.deep_merge(attributes)
+        def merchant_account_token(id, members)
+          t = Shore::Client::Tokens::V1::MerchantAccount.new(id)
+          t.import_roles(members)
+          t
         end
       end
     end
